arm64: dts: rockchip: Add RTL8367RB-VB switch to Radxa E24C
The Radxa E24C has a Realtek RTL8367RB-VB switch with four usable ports
and is connected using a fixed-link to GMAC1 on the RK3528 SoC.

TSO and/or RX checksum offload may need to be disabled on the conduit
interface (gmac1).

Add an ethernet-switch node to describe the RTL8367RB-VB switch.

Signed-off-by: Jonas Karlman <jonas@kwiboo.se>